Reformatting the preferences activity to free up space in MainActivity toolbar.
