+++ app/src/main/res/drawable/preference_button.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="@color/grey" />
+        </shape>
+    </item>
+    <item android:bottom="0.8dp">
+        <shape android:shape="rectangle">
+            <solid android:color="@color/background_material_dark" />
+        </shape>
+    </item>
+</layer-list>
